Change in tip 1 (Lesson 5) and New challenge for lesson 01
Change in tip 1 (Lesson 5) to try to improve its usefulness

Currently the tip is worded in a way that might mislead novices, as they
might think that the only reason for failure is the one discussed in the
tip.
I have included part of the error message in the tip, so that it is
easier to know if the solution provided is adequate and will work.

New challenge for lesson 01

Many participants can have data that can with different formats from the
standard csv described. The objective of the challenge is to make
participants aware of this potential problem, encourage them to use the
help functions available in R, and to practice with a simple coding
exercise early on.