add alt text for multiple plots
This one in particular was tricky because all of these plots are very similar and the point of the exercise was to show that you could run a plotting function over a set of files as opposed to running them by hand. Thus, just as a sighted user would skim over the plots, a user who relies on a screen reader would do so as well.